feat(project): refactor structure and namings
